feat: error handling

+++ backend/src/presentation/controllers/agency.controller.ts
@@ -6,11 +6,13 @@ import {
   Delete,
   Param,
   Body,
+  HttpException,
 } from '@nestjs/common';
 import { AgencyRepository } from '@/infrastructure/repository/Agency/agency.repository';
 import { CreateAgencyDto } from '@/presentation/dtos/Agency/create-agency-dto';
 import type { UpdateDtoFor } from '@/presentation/dtos/.Base/base-dtos';
 import { Agency } from '@/domain/entities/Agency/agency.schema';
+import { Types } from 'mongoose';
 
 @Controller('agency')
 export class AgencyController {
@@ -23,7 +25,11 @@ export class AgencyController {
 
   @Get(':id')
   async findById(@Param('id') id: string) {
-    return this.agencyRepository.findById(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Agency not found', 404);
+    const findAgency = await this.agencyRepository.findById(id);
+    if (!findAgency) throw new HttpException('Agency not found', 404);
+    return findAgency;
   }
 
   @Post()
@@ -36,11 +42,19 @@ export class AgencyController {
     @Param('id') id: string,
     @Body() updateDto: UpdateDtoFor<Agency>,
   ) {
-    return this.agencyRepository.update(id, updateDto);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const updatedAgency = await this.agencyRepository.update(id, updateDto);
+    if (!updatedAgency) throw new HttpException('Agency Not Found', 404);
+    return updatedAgency;
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
-    return this.agencyRepository.delete(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const deletedAgency = await this.agencyRepository.delete(id);
+    if (!deletedAgency) throw new HttpException('Agency Not Found', 404);
+    return deletedAgency;
   }
 }
