+++ backend/src/presentation/controllers/agent.controller.ts
@@ -11,7 +11,8 @@ import { AgentRepository } from '@/infrastructure/repository/Agent/agent.reposit
 import { CreateAgentDto } from '@/presentation/dtos/Agent/create-agent-dto';
 import type { UpdateDtoFor } from '@/presentation/dtos/.Base/base-dtos';
 import { Agent } from '@/domain/entities/Agent/agent.schema';
-
+import { Types } from 'mongoose';
+import { HttpException } from '@nestjs/common';
 @Controller('agents')
 export class AgentController {
   constructor(private readonly agentRepository: AgentRepository) {}
@@ -23,7 +24,11 @@ export class AgentController {
 
   @Get(':id')
   async findById(@Param('id') id: string) {
-    return this.agentRepository.findById(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Agent not found', 404);
+    const findAgent = await this.agentRepository.findById(id);
+    if (!findAgent) throw new HttpException('Agent not found', 404);
+    return findAgent;
   }
 
   @Post()
@@ -36,11 +41,19 @@ export class AgentController {
     @Param('id') id: string,
     @Body() updateDto: UpdateDtoFor<Agent>,
   ) {
-    return this.agentRepository.update(id, updateDto);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const updatedAgent = await this.agentRepository.update(id, updateDto);
+    if (!updatedAgent) throw new HttpException('Agent Not Found', 404);
+    return updatedAgent;
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
-    return this.agentRepository.delete(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const deletedAgent = await this.agentRepository.delete(id);
+    if (!deletedAgent) throw new HttpException('Agent Not Found', 404);
+    return deletedAgent;
   }
 }
