+++ backend/src/presentation/controllers/property.controller.ts
@@ -6,11 +6,13 @@ import {
   Delete,
   Param,
   Body,
+  HttpException,
 } from '@nestjs/common';
 import { PropertyRepository } from '@/infrastructure/repository/Property/property.repository';
 import { CreatePropertyDto } from '@/presentation/dtos/Property/create-property-dto';
 import type { UpdateDtoFor } from '@/presentation/dtos/.Base/base-dtos';
 import { Property } from '@/domain/entities/Property/property.schema';
+import { Types } from 'mongoose';
 
 @Controller('properties')
 export class PropertyController {
@@ -23,7 +25,11 @@ export class PropertyController {
 
   @Get(':id')
   async findById(@Param('id') id: string) {
-    return this.propertyRepository.findById(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Property not found', 404);
+    const findProperty = await this.propertyRepository.findById(id);
+    if (!findProperty) throw new HttpException('Property not found', 404);
+    return findProperty;
   }
 
   @Post()
@@ -36,11 +42,19 @@ export class PropertyController {
     @Param('id') id: string,
     @Body() updateDto: UpdateDtoFor<Property>,
   ) {
-    return this.propertyRepository.update(id, updateDto);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const updatedProperty = await this.propertyRepository.update(id, updateDto);
+    if (!updatedProperty) throw new HttpException('Property Not Found', 404);
+    return updatedProperty;
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
-    return this.propertyRepository.delete(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const deletedProperty = await this.propertyRepository.delete(id);
+    if (!deletedProperty) throw new HttpException('Property Not Found', 404);
+    return deletedProperty;
   }
 }
