+++ backend/src/presentation/controllers/transaction.controller.ts
@@ -6,7 +6,7 @@ import {
   Delete,
   Param,
   Body,
-  NotFoundException,
+  HttpException,
 } from '@nestjs/common';
 import { TransactionRepository } from '@/infrastructure/repository/Transaction/transaction.repository';
 import { CreateTransactionDto } from '@/presentation/dtos/Transaction/create-transaction-dto';
@@ -14,6 +14,8 @@ import type { UpdateDtoFor } from '@/presentation/dtos/.Base/base-dtos';
 import { Transaction } from '@/domain/entities/Transaction/transaction.schema';
 import { UpdateTransactionStageUseCase } from '@/application/use-cases/Transaction/updateTransactionStage.use-case';
 import { GetFinancialBreakdownUseCase } from '@/application/use-cases/Transaction/getFinancialBreakdown.use-case';
+import { Types } from 'mongoose';
+
 @Controller('transactions')
 export class TransactionController {
   constructor(
@@ -29,14 +31,20 @@ export class TransactionController {
 
   @Get(':id')
   async findById(@Param('id') id: string) {
-    return this.transactionRepository.findById(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Transaction not found', 404);
+    const findTransaction = await this.transactionRepository.findById(id);
+    if (!findTransaction) throw new HttpException('Transaction not found', 404);
+    return findTransaction;
   }
 
   @Get(':id/breakdown')
   async getBreakdown(@Param('id') id: string) {
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Transaction not found', 404);
     const transaction = await this.transactionRepository.findById(id);
     if (!transaction) {
-      throw new NotFoundException('Transaction not found');
+      throw new HttpException('Transaction not found', 404);
     }
     return this.getBreakdownUseCase.execute(transaction);
   }
@@ -51,14 +59,25 @@ export class TransactionController {
     @Param('id') id: string,
     @Body() updateDto: UpdateDtoFor<Transaction>,
   ) {
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+
     if (updateDto.stage) {
-      return this.updateStageUseCase.execute(id, updateDto.stage);
+      const updatedTransaction = await this.updateStageUseCase.execute(id, updateDto.stage);
+      if (!updatedTransaction) throw new HttpException('Transaction Not Found', 404);
+      return updatedTransaction;
     }
-    return this.transactionRepository.update(id, updateDto);
+    const updatedTransaction = await this.transactionRepository.update(id, updateDto);
+    if (!updatedTransaction) throw new HttpException('Transaction Not Found', 404);
+    return updatedTransaction;
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string) {
-    return this.transactionRepository.delete(id);
+    const isValid = Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const deletedTransaction = await this.transactionRepository.delete(id);
+    if (!deletedTransaction) throw new HttpException('Transaction Not Found', 404);
+    return deletedTransaction;
   }
 }
